Cleaned up the 'Constant' module
