+++ language-plutus-core/src/Language/PlutusCore/Constant.hs
@@ -4,7 +4,10 @@ module Language.PlutusCore.Constant
     ( module Export
     ) where
 
-import           Language.PlutusCore.Constant.Apply       as Export
-import           Language.PlutusCore.Constant.DynamicType as Export
-import           Language.PlutusCore.Constant.Make        as Export
-import           Language.PlutusCore.Constant.Typed       as Export
+import           Language.PlutusCore.Constant.Apply            as Export
+import           Language.PlutusCore.Constant.DynamicInstances as Export ()
+import           Language.PlutusCore.Constant.Function         as Export
+import           Language.PlutusCore.Constant.Make             as Export
+import           Language.PlutusCore.Constant.Name             as Export
+import           Language.PlutusCore.Constant.PrettyDynamic    as Export
+import           Language.PlutusCore.Constant.Typed            as Export
