+++ language-plutus-core/src/Language/PlutusCore/Constant/Function.hs
@@ -1,0 +1,159 @@
+{-# LANGUAGE GADTs             #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE RankNTypes        #-}
+
+module Language.PlutusCore.Constant.Function
+    ( flattenSizeEntry
+    , eraseTypedBuiltinSized
+    , mapSizeEntryTypedBuiltin
+    , mapSizeTypedBuiltin
+    , closeTypedBuiltin
+    , typedBuiltinSizedToType
+    , withTypedBuiltinSized
+    , withTypedBuiltin
+    , typeSchemeResult
+    , typedBuiltinToType
+    , typeSchemeToType
+    , dynamicBuiltinNameMeaningToType
+    , insertDynamicBuiltinNameDefinition
+    , unionDynamicBuiltinNameMeanings
+    , withTypedBuiltinName
+    , typeOfTypedBuiltinName
+    , typeOfBuiltinName
+    ) where
+
+import           Language.PlutusCore.Constant.Name
+import           Language.PlutusCore.Constant.Typed
+import           Language.PlutusCore.Lexer.Type       hiding (name)
+import           Language.PlutusCore.Name
+import           Language.PlutusCore.Pretty
+import           Language.PlutusCore.Quote
+import           Language.PlutusCore.StdLib.Data.Bool
+import           Language.PlutusCore.Type
+
+import qualified Data.ByteString.Lazy.Char8           as BSL
+import qualified Data.Map                             as Map
+import qualified Data.Text.Encoding                   as Text
+
+-- | Extract the 'Size' from a 'SizeEntry'.
+flattenSizeEntry :: SizeEntry Size -> Size
+flattenSizeEntry (SizeValue size) = size
+flattenSizeEntry (SizeBound size) = size
+
+-- | Alter the 'SizeEntry' of a 'TypedBuiltin'.
+mapSizeEntryTypedBuiltin
+    :: (SizeEntry size -> SizeEntry size') -> TypedBuiltin size a -> TypedBuiltin size' a
+mapSizeEntryTypedBuiltin f (TypedBuiltinSized se tbs) = TypedBuiltinSized (f se) tbs
+mapSizeEntryTypedBuiltin _ TypedBuiltinBool           = TypedBuiltinBool
+mapSizeEntryTypedBuiltin _ TypedBuiltinDyn            = TypedBuiltinDyn
+
+-- | Alter the 'size' of a @TypedBuiltin size@.
+mapSizeTypedBuiltin
+    :: (size -> size') -> TypedBuiltin size a -> TypedBuiltin size' a
+mapSizeTypedBuiltin = mapSizeEntryTypedBuiltin . fmap
+
+-- | Map each 'SizeBound' to 'SizeValue'.
+closeTypedBuiltin :: TypedBuiltin Size a -> TypedBuiltin b a
+closeTypedBuiltin = mapSizeEntryTypedBuiltin $ SizeValue . flattenSizeEntry
+
+-- | Convert a 'TypedBuiltinSized' to the corresponding 'TypeBuiltin' and
+-- wrap the result in 'TyBuiltin' to get a 'Type'.
+typedBuiltinSizedToType :: TypedBuiltinSized a -> Type TyName ()
+typedBuiltinSizedToType TypedBuiltinSizedInt  = TyBuiltin () TyInteger
+typedBuiltinSizedToType TypedBuiltinSizedBS   = TyBuiltin () TyByteString
+typedBuiltinSizedToType TypedBuiltinSizedSize = TyBuiltin () TySize
+
+-- | Apply a continuation to the typed version of a 'BuiltinSized'.
+withTypedBuiltinSized :: BuiltinSized -> (forall a. TypedBuiltinSized a -> c) -> c
+withTypedBuiltinSized BuiltinSizedInt  k = k TypedBuiltinSizedInt
+withTypedBuiltinSized BuiltinSizedBS   k = k TypedBuiltinSizedBS
+withTypedBuiltinSized BuiltinSizedSize k = k TypedBuiltinSizedSize
+
+-- | Apply a continuation to the typed version of a 'Builtin'.
+withTypedBuiltin :: BuiltinType size -> (forall a. TypedBuiltin size a -> c) -> c
+withTypedBuiltin (BuiltinSized se b) k = withTypedBuiltinSized b $ k . TypedBuiltinSized se
+withTypedBuiltin BuiltinBool         k = k TypedBuiltinBool
+
+-- | The resulting 'TypedBuiltin' of a 'TypeScheme'.
+typeSchemeResult :: TypeScheme () a r -> TypedBuiltin () r
+typeSchemeResult (TypeSchemeBuiltin tb)   = tb
+typeSchemeResult (TypeSchemeArrow _ schB) = typeSchemeResult schB
+typeSchemeResult (TypeSchemeAllSize schK) = typeSchemeResult (schK ())
+
+-- | Convert a 'TypedBuiltin' to the corresponding 'Type'.
+typedBuiltinToType :: TypedBuiltin (Type TyName ()) a -> Quote (Type TyName ())
+typedBuiltinToType (TypedBuiltinSized se tbs) =
+    return . TyApp () (typedBuiltinSizedToType tbs) $ case se of
+        SizeValue size -> TyInt () size
+        SizeBound ty   -> ty
+typedBuiltinToType TypedBuiltinBool           = getBuiltinBool
+typedBuiltinToType dyn@TypedBuiltinDyn        = getTypeEncoding dyn
+
+-- | Convert a 'TypeScheme' to the corresponding 'Type'.
+-- Basically, a map from the PHOAS representation to the FOAS one.
+typeSchemeToType :: TypeScheme (Type TyName ()) a r -> Quote (Type TyName ())
+typeSchemeToType = go 0 where
+    go :: Int -> TypeScheme (Type TyName ()) a r -> Quote (Type TyName ())
+    go _ (TypeSchemeBuiltin tb)      = typedBuiltinToType tb
+    go i (TypeSchemeArrow schA schB) =
+        TyFun () <$> go i schA <*> go i schB
+    go i (TypeSchemeAllSize schK)    = do
+        s <- mapTyNameString (<> BSL.fromStrict (Text.encodeUtf8 $ prettyText i)) <$>
+                freshTyName () "s"
+        a <- go (succ i) . schK $ TyVar () s
+        return $ TyForall () s (Size ()) a
+
+-- | Extract the 'TypeScheme' from a 'DynamicBuiltinNameMeaning' and
+-- convert it to the corresponding 'Type'.
+dynamicBuiltinNameMeaningToType :: DynamicBuiltinNameMeaning -> Quote (Type TyName ())
+dynamicBuiltinNameMeaningToType (DynamicBuiltinNameMeaning sch _) = typeSchemeToType sch
+
+-- | Insert a 'DynamicBuiltinNameDefinition' into a 'DynamicBuiltinNameMeanings'.
+insertDynamicBuiltinNameDefinition
+    :: DynamicBuiltinNameDefinition -> DynamicBuiltinNameMeanings -> DynamicBuiltinNameMeanings
+insertDynamicBuiltinNameDefinition
+    (DynamicBuiltinNameDefinition name mean) (DynamicBuiltinNameMeanings nameMeans) =
+        DynamicBuiltinNameMeanings $ Map.insert name mean nameMeans
+
+-- | Union two 'DynamicBuiltinNameMeanings's.
+unionDynamicBuiltinNameMeanings
+    :: DynamicBuiltinNameMeanings -> DynamicBuiltinNameMeanings -> DynamicBuiltinNameMeanings
+unionDynamicBuiltinNameMeanings
+    (DynamicBuiltinNameMeanings means1) (DynamicBuiltinNameMeanings means2) =
+        DynamicBuiltinNameMeanings $ Map.union means1 means2
+
+-- | Apply a continuation to the typed version of a 'BuiltinName'.
+withTypedBuiltinName :: BuiltinName -> (forall a r. TypedBuiltinName a r -> c) -> c
+withTypedBuiltinName AddInteger           k = k typedAddInteger
+withTypedBuiltinName SubtractInteger      k = k typedSubtractInteger
+withTypedBuiltinName MultiplyInteger      k = k typedMultiplyInteger
+withTypedBuiltinName DivideInteger        k = k typedDivideInteger
+withTypedBuiltinName QuotientInteger      k = k typedQuotientInteger
+withTypedBuiltinName RemainderInteger     k = k typedRemainderInteger
+withTypedBuiltinName ModInteger           k = k typedModInteger
+withTypedBuiltinName LessThanInteger      k = k typedLessThanInteger
+withTypedBuiltinName LessThanEqInteger    k = k typedLessThanEqInteger
+withTypedBuiltinName GreaterThanInteger   k = k typedGreaterThanInteger
+withTypedBuiltinName GreaterThanEqInteger k = k typedGreaterThanEqInteger
+withTypedBuiltinName EqInteger            k = k typedEqInteger
+withTypedBuiltinName ResizeInteger        k = k typedResizeInteger
+withTypedBuiltinName IntToByteString      k = k typedIntToByteString
+withTypedBuiltinName Concatenate          k = k typedConcatenate
+withTypedBuiltinName TakeByteString       k = k typedTakeByteString
+withTypedBuiltinName DropByteString       k = k typedDropByteString
+withTypedBuiltinName SHA2                 k = k typedSHA2
+withTypedBuiltinName SHA3                 k = k typedSHA3
+withTypedBuiltinName VerifySignature      k = k typedVerifySignature
+withTypedBuiltinName ResizeByteString     k = k typedResizeByteString
+withTypedBuiltinName EqByteString         k = k typedEqByteString
+withTypedBuiltinName TxHash               k = k typedTxHash
+withTypedBuiltinName BlockNum             k = k typedBlockNum
+withTypedBuiltinName SizeOfInteger        k = k typedSizeOfInteger
+
+-- | Return the 'Type' of a 'TypedBuiltinName'.
+typeOfTypedBuiltinName :: TypedBuiltinName a r -> Quote (Type TyName ())
+typeOfTypedBuiltinName (TypedBuiltinName _ scheme) = typeSchemeToType scheme
+
+-- | Return the 'Type' of a 'BuiltinName'.
+typeOfBuiltinName :: BuiltinName -> Quote (Type TyName ())
+typeOfBuiltinName bn = withTypedBuiltinName bn typeOfTypedBuiltinName
