+++ language-plutus-core/src/Language/PlutusCore/Constant/Name.hs
@@ -1,0 +1,205 @@
+module Language.PlutusCore.Constant.Name
+    ( typedAddInteger
+    , typedSubtractInteger
+    , typedMultiplyInteger
+    , typedDivideInteger
+    , typedQuotientInteger
+    , typedModInteger
+    , typedRemainderInteger
+    , typedLessThanInteger
+    , typedLessThanEqInteger
+    , typedGreaterThanInteger
+    , typedGreaterThanEqInteger
+    , typedEqInteger
+    , typedResizeInteger
+    , typedIntToByteString
+    , typedConcatenate
+    , typedTakeByteString
+    , typedDropByteString
+    , typedSHA2
+    , typedSHA3
+    , typedVerifySignature
+    , typedResizeByteString
+    , typedEqByteString
+    , typedTxHash
+    , typedBlockNum
+    , typedSizeOfInteger
+    ) where
+
+import           Language.PlutusCore.Constant.Typed
+import           Language.PlutusCore.Lexer.Type
+
+import qualified Data.ByteString.Lazy.Char8         as BSL
+
+sizeIntIntInt :: TypeScheme size (Integer -> Integer -> Integer) Integer
+sizeIntIntInt =
+    TypeSchemeAllSize $ \s ->
+        TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedInt) `TypeSchemeArrow`
+        TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedInt) `TypeSchemeArrow`
+        TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedInt)
+
+sizeIntIntBool :: TypeScheme size (Integer -> Integer -> Bool) Bool
+sizeIntIntBool =
+    TypeSchemeAllSize $ \s ->
+        TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedInt) `TypeSchemeArrow`
+        TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedInt) `TypeSchemeArrow`
+        TypeSchemeBuiltin TypedBuiltinBool
+
+-- | Typed 'AddInteger'.
+typedAddInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer
+typedAddInteger = TypedBuiltinName AddInteger sizeIntIntInt
+
+-- | Typed 'SubtractInteger'.
+typedSubtractInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer
+typedSubtractInteger = TypedBuiltinName SubtractInteger sizeIntIntInt
+
+-- | Typed 'MultiplyInteger'.
+typedMultiplyInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer
+typedMultiplyInteger = TypedBuiltinName MultiplyInteger sizeIntIntInt
+
+-- | Typed 'DivideInteger'.
+typedDivideInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer
+typedDivideInteger = TypedBuiltinName DivideInteger sizeIntIntInt
+
+-- | Typed 'QuotientInteger'
+typedQuotientInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer
+typedQuotientInteger = TypedBuiltinName QuotientInteger sizeIntIntInt
+
+-- | Typed 'RemainderInteger'.
+typedRemainderInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer
+typedRemainderInteger = TypedBuiltinName RemainderInteger sizeIntIntInt
+
+-- | Typed 'ModInteger'
+typedModInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer
+typedModInteger = TypedBuiltinName ModInteger sizeIntIntInt
+
+-- | Typed 'LessThanInteger'.
+typedLessThanInteger :: TypedBuiltinName (Integer -> Integer -> Bool) Bool
+typedLessThanInteger = TypedBuiltinName LessThanInteger sizeIntIntBool
+
+-- | Typed 'LessThanEqInteger'.
+typedLessThanEqInteger :: TypedBuiltinName (Integer -> Integer -> Bool) Bool
+typedLessThanEqInteger = TypedBuiltinName LessThanEqInteger sizeIntIntBool
+
+-- | Typed 'GreaterThanInteger'.
+typedGreaterThanInteger :: TypedBuiltinName (Integer -> Integer -> Bool) Bool
+typedGreaterThanInteger = TypedBuiltinName GreaterThanInteger sizeIntIntBool
+
+-- | Typed 'GreaterThanEqInteger'.
+typedGreaterThanEqInteger :: TypedBuiltinName (Integer -> Integer -> Bool) Bool
+typedGreaterThanEqInteger = TypedBuiltinName GreaterThanEqInteger sizeIntIntBool
+
+-- | Typed 'EqInteger'.
+typedEqInteger :: TypedBuiltinName (Integer -> Integer -> Bool) Bool
+typedEqInteger = TypedBuiltinName EqInteger sizeIntIntBool
+
+-- | Typed 'ResizeInteger'.
+typedResizeInteger :: TypedBuiltinName (() -> Integer -> Integer) Integer
+typedResizeInteger =
+    TypedBuiltinName ResizeInteger $
+        TypeSchemeAllSize $ \s0 -> TypeSchemeAllSize $ \s1 ->
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s1) TypedBuiltinSizedSize) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s0) TypedBuiltinSizedInt) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s1) TypedBuiltinSizedInt)
+
+-- | Typed 'IntToByteString'.
+typedIntToByteString :: TypedBuiltinName (() -> Integer -> BSL.ByteString) BSL.ByteString
+typedIntToByteString =
+    TypedBuiltinName IntToByteString $
+        TypeSchemeAllSize $ \s0 -> TypeSchemeAllSize $ \s1 ->
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s1) TypedBuiltinSizedSize) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s0) TypedBuiltinSizedInt) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s1) TypedBuiltinSizedBS)
+
+-- | Typed 'Concatenate'.
+typedConcatenate :: TypedBuiltinName (BSL.ByteString -> BSL.ByteString -> BSL.ByteString) BSL.ByteString
+typedConcatenate =
+    TypedBuiltinName Concatenate $
+        TypeSchemeAllSize $ \s ->
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedBS) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedBS) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedBS)
+
+-- | Typed 'TakeByteString'.
+typedTakeByteString :: TypedBuiltinName (Integer -> BSL.ByteString -> BSL.ByteString) BSL.ByteString
+typedTakeByteString =
+    TypedBuiltinName TakeByteString $
+        TypeSchemeAllSize $ \s0 -> TypeSchemeAllSize $ \s1 ->
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s0) TypedBuiltinSizedInt) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s1) TypedBuiltinSizedBS) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s1) TypedBuiltinSizedBS)
+
+-- | Typed 'DropByteString'.
+typedDropByteString :: TypedBuiltinName (Integer -> BSL.ByteString -> BSL.ByteString) BSL.ByteString
+typedDropByteString =
+    TypedBuiltinName DropByteString $
+        TypeSchemeAllSize $ \s0 -> TypeSchemeAllSize $ \s1 ->
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s0) TypedBuiltinSizedInt) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s1) TypedBuiltinSizedBS) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s1) TypedBuiltinSizedBS)
+
+-- | Typed 'SHA2'.
+typedSHA2 :: TypedBuiltinName (BSL.ByteString -> BSL.ByteString) BSL.ByteString
+typedSHA2 =
+    TypedBuiltinName SHA2 $
+        TypeSchemeAllSize $ \s ->
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedBS) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeValue 256) TypedBuiltinSizedBS)
+
+-- | Typed 'SHA3'.
+typedSHA3 :: TypedBuiltinName (BSL.ByteString -> BSL.ByteString) BSL.ByteString
+typedSHA3 =
+    TypedBuiltinName SHA3 $
+        TypeSchemeAllSize $ \s ->
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedBS) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeValue 256) TypedBuiltinSizedBS)
+
+-- | Typed 'VerifySignature'.
+typedVerifySignature :: TypedBuiltinName (BSL.ByteString -> BSL.ByteString -> BSL.ByteString -> Bool) Bool
+typedVerifySignature =
+    TypedBuiltinName VerifySignature $
+        TypeSchemeAllSize $ \s0 -> TypeSchemeAllSize $ \s1 -> TypeSchemeAllSize $ \s2 ->
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s0) TypedBuiltinSizedBS) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s1) TypedBuiltinSizedBS) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s2) TypedBuiltinSizedBS) `TypeSchemeArrow`
+            TypeSchemeBuiltin TypedBuiltinBool
+
+-- | Typed 'ResizeByteString'.
+typedResizeByteString :: TypedBuiltinName (() -> BSL.ByteString -> BSL.ByteString) BSL.ByteString
+typedResizeByteString =
+    TypedBuiltinName ResizeByteString $
+        TypeSchemeAllSize $ \s0 -> TypeSchemeAllSize $ \s1 ->
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s1) TypedBuiltinSizedSize) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s0) TypedBuiltinSizedBS) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s1) TypedBuiltinSizedBS)
+
+-- | Typed 'EqByteString'.
+typedEqByteString :: TypedBuiltinName (BSL.ByteString -> BSL.ByteString -> Bool) Bool
+typedEqByteString =
+    TypedBuiltinName EqByteString $
+        TypeSchemeAllSize $ \s ->
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedBS) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedBS) `TypeSchemeArrow`
+            TypeSchemeBuiltin TypedBuiltinBool
+
+-- | Typed 'TxHash'.
+typedTxHash :: TypedBuiltinName BSL.ByteString BSL.ByteString
+typedTxHash =
+    TypedBuiltinName TxHash $
+        TypeSchemeBuiltin (TypedBuiltinSized (SizeValue 256) TypedBuiltinSizedBS)
+
+-- | Typed 'BlockNum'.
+typedBlockNum :: TypedBuiltinName (() -> Integer) Integer
+typedBlockNum =
+    TypedBuiltinName BlockNum $
+        TypeSchemeAllSize $ \s ->
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedSize) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedInt)
+
+-- | Typed 'SizeOfInteger'.
+typedSizeOfInteger :: TypedBuiltinName (Integer -> ()) ()
+typedSizeOfInteger =
+    TypedBuiltinName SizeOfInteger $
+        TypeSchemeAllSize $ \s ->
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedInt) `TypeSchemeArrow`
+            TypeSchemeBuiltin (TypedBuiltinSized (SizeBound s) TypedBuiltinSizedSize)
